requirements.txt
numpy
pandas
streamlit
matplotlib
openai>=1.0.0
python-dotenv
